Refactor: python, cts: Move ResourceRecover into the _cts module.
No code changes.
