+++ python/pacemaker/_cts/tests/__init__.py
@@ -10,6 +10,7 @@ from pacemaker._cts.tests.fliptest import FlipTest
 from pacemaker._cts.tests.maintenancemode import MaintenanceMode
 from pacemaker._cts.tests.partialstart import PartialStart
 from pacemaker._cts.tests.restartonebyone import RestartOnebyOne
+from pacemaker._cts.tests.resourcerecover import ResourceRecover
 from pacemaker._cts.tests.restarttest import RestartTest
 from pacemaker._cts.tests.remotedriver import RemoteDriver
 from pacemaker._cts.tests.simulstart import SimulStart
